Доработан движок: осуществлён переход на асинхронные операции
Доработан движок TVProgViewer:
Осуществлён переход на асинхронные операции.

+++ src/TVProgCoreMvc/TVProgViewer.Core/Domain/TvProgMain/ExtUserSettings.cs
@@ -18,6 +18,6 @@
         /// <summary>
         /// Неотмеченные телеканалы
         /// </summary>
-        public short? UncheckedChannels { get; set; }
+        public int? UncheckedChannels { get; set; }
     }
 }
